+++ Functionality to add new elements to a list
Take in consideration that it only adds elements to the "Upcoming vaccines" list, and only adds the specified text, but not the date. Next step is to adapt the method so it can add new elements to both vaccine and prescription lists, and to add the date to the elements so the list can be ordered.